add bigquery integration test shim (#16)
Just realized that the BigQuery integration tests when the
`fullstory_events_integration_tests` didn't previously exist. This
creates a BigQuery shim that is very similar to the Snowflake one except
that this one doesn't do any JSON parsing, it just takes the seed file
as is.

This also makes tests for each platform all work the same way to
increase clarity.