fix incremental model, add tests (#39)
This fixes a syntax error that was uncovered when running in incremental
mode 🤦. I added a provision to the automated tests that will now run dbt
twice, once with full refresh and once in incremental mode.

I also changed to use the dbt.cast macro and dbt.dateadd macro to make
things work properly across databases (uncovered with the new tests).