v0.9.2 - allow nullable updated_time in events (#43)
# dbt_fullstory Changelog

## [0.9.2] - 2025-07-21
- **Coalesce updated_time in staging model** - Fullstory data stored in
object storage (i.e. S3, GCS, or Azure Blobk Storage), may not have an
`updated_time` value and is nullable on those destinations. Therefore
the `updated_time` column will be coalesced with `processed_time` within
the `stg_fullstory__events` model.

---

This pull request introduces changes to enhance data handling in the
`stg_fullstory__events` model and updates the project version to reflect
the modifications. The most significant updates include coalescing the
`updated_time` column with `processed_time` to handle nullable values,
and updating the project version and changelog accordingly.

### Enhancements to data handling:

*
[`models/staging/stg_fullstory__events.sql`](diffhunk://#diff-1cdc2ed718eaaad3058fbb4107e4ddc3bbf5f3eee552c80b25b1635393cfbf93L16-R19):
Modified the `updated_time` column to use `coalesce` with
`processed_time` to ensure a fallback value when `updated_time` is null.
This improves data consistency for Fullstory data stored in object
storage.

### Project versioning and documentation:

*
[`dbt_project.yml`](diffhunk://#diff-27386df54b2629c1191d8342d3725ed8678413cfa13b5556f59d69d33fae5425L2-R2):
Updated the project version from `0.9.0` to `0.9.2` to reflect the new
changes.
*
[`CHANGELOG.md`](diffhunk://#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4edR8-R10):
Added an entry for version `0.9.2` detailing the coalescing of
`updated_time` in the `stg_fullstory__events` model.

---------

Co-authored-by: Trevor Hutto <trevorhutto@fullstory.com>