#45 - macro call in wrong spot (#46)
This pull request makes a small but important change to the filtering
logic in the `models/events.sql` file. The placement of the `where
event_id_rn = 1` clause has been adjusted to ensure correct query
execution and incremental logic.

- SQL logic update:
* Moved the `where event_id_rn = 1` clause to appear before the
incremental filters, ensuring that the row number filter is always
applied regardless of whether the query is running incrementally or as a
full refresh.